requirements.txt
python-box
pyyaml